+++ opencl/src/fem/structures.rs
@@ -1,5 +1,5 @@
 use nalgebra::{DMatrix, DVector, Dyn};
 
-pub type Matrix = DMatrix<f32>;
-pub type LU = nalgebra::LU<f32, Dyn, Dyn>;
-pub type Vector = DVector<f32>;
+pub type Matrix = DMatrix<f64>;
+pub type LU = nalgebra::LU<f64, Dyn, Dyn>;
+pub type Vector = DVector<f64>;
